添加了baseTest方法，把根据id查询的类型换成了Integer类型

+++ src/main/java/com/p2p/service/impl/BorrowApplyServiceImpl.java
@@ -50,7 +50,7 @@ public class BorrowApplyServiceImpl implements BorrowApplyService {
     }
 
     @Override
-    public Object getById(Long id) {
+    public Object getById(Integer id) {
         return null;
     }
